Fix kyverno policy (#173)
* The kyverno policy does not work on OpenShift

The kyverno helm chart had two main problems when installing it on
OpenShift:
1. The default memory limit of 256Mi was too low.  I've seen the limit
even exceed 512 briefly so using 768Mi as the current limit.  I have
not done performance testing or tried to determine why the usage may
be high on OpenShift.  I suspect it's because OpenShift has many more
namespaces and other resources than some other environments.
2. When installing kyverno on the ACM Hub, there's a conflict with the
`policyreport` CRD which already exists due to insights.  The fix is
to make kyverno think it owns the CRD.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/16954

Signed-off-by: Gus Parvin <gparvin@redhat.com>

* Remove hard coded namespace value

The policy should not specify the namespace.

Signed-off-by: Gus Parvin <gparvin@redhat.com>