Remove starting csv (#186)
* Remove some hard coded versions or update tags

We don't always need the CSV version to be hard coded. Removing some
of those and using policies that don't specify image versions
when possible.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/17105

Signed-off-by: Gus Parvin <gparvin@redhat.com>

* Remove CSV and update channel for policies that install operators

Many policies specify a starting CSV version when an operator is
installed.  This can be removed.  Sometimes the channel is based on
the OpenShift version, which can be updated or changed to stable
when possible.

Refs:
 - https://github.com/open-cluster-management/backlog/issues/17105

Signed-off-by: Gus Parvin <gparvin@redhat.com>

* Not removing the odf policy as part of this PR

Handle removing the odf policy separately if necessary

Signed-off-by: Gus Parvin <gparvin@redhat.com>