Fix problems with ODF and eliminate need for scripts (#229)
There were some problems with the ODF install.  The operator now
installs and the storagesystem and the storagecluster are properly
setup.  Observability I think is still have trouble writing to noobaa,
but there are statistics and details that show things are closer to
working properly.  While I made those changes I also eliminated the
scripts required by taking Andrew's advice and incorporating the post
scripts as jobs that run in the local-quay and stackrox namespaces
to setup secrets those apps need.

Signed-off-by: Gus Parvin <gparvin@redhat.com>