Wrong namespace used for new managed cluster hub templates
The ACS Sensor secrets were pulling from the wrong namespace -- likely
a copy/paste issue from moving code around to support the job that
creates the secrets a little differently.  It worked fine on the hub,
but creating a managed cluster showed the issue.

I also had to move the SecuredCluster resource so that it was after
the secrets.  The SecuredCluster errored the policy out so the
secrets were never created, and they need to be there first.

Refs:
 - https://github.com/stolostron/backlog/issues/20093

Signed-off-by: Gus Parvin <gparvin@redhat.com>