Create the kyverno policies to the community that use generate (#232)
The Kyverno generate policies are a unique set of policies that can
fill gaps around `NetworkPolicy`, quotas and syncing secrets in a
unique and valuable way.  We would like these policies to become
stable, but first, let's add them to the community.

Note: my previous pr was accidentally wiped out when I deleted my fork.
See the details in that PR:
- https://github.com/stolostron/policy-collection/pull/228

Signed-off-by: Gus Parvin <gparvin@redhat.com>