This is a MachineSets policy that is useful for OCS on AWS (#246)
To install OCS on an AWS cluster you need some extra `MachineSets`
that will be used by OCS.  This policy shows how to use templates to
easily create the MachineSets with the cluster name and region
configured using templates.  The AMI ID does have to be updated to
a proper value for your AWS region and OpenShift version.

Signed-off-by: Gus Parvin <gparvin@redhat.com>