Address problem downloading the cli image in the OPP policyset
The openshift CLI image sometimes can't be downloaded.  In one scenario
I recreated the issue and noticed one pod successfully downloaded the
image and another pod didn't.  Both pods were on the same node so I
deleted the pod that failed to download and when the pod restarted it
worked find.  My guess is there may be a permissions or timing issue
that caused the problem so I have re-ordered some resources to change
the timing.

Currently this is a guess so it may turn into some job configuration.

Signed-off-by: Gus Parvin <gparvin@redhat.com>