Some kyverno policies should not do background checking (#346)
Any kyverno policy that is working with cluster scoped resources will
potentially match lots of resources if `background: true` is set in the
ClusterPolicy.  Example: any of the policies that look at ClusterRoles.
If the resource is namespaced we do some namespace filtering to keep
from matching the infrastructure resources -- but we can't do that
filtering for all cluster scoped resources (maybe we could do some --
but prefer to just turn background off)

Refs:
 - https://github.com/stolostron/backlog/issues/25864

Signed-off-by: Gus Parvin <gparvin@redhat.com>

Signed-off-by: Gus Parvin <gparvin@redhat.com>