Links and timing concerns for quay
This commit adds the links for:
- acm observability grafana
- ACS
- quay

This commit re-arranges the quay and ACS resources.  The ACS resources
that contain templates are now separated into a sync resources policy
which will need to always have `consolidateManifests` set to false.
The quay hub install now has an install and a config policy.  The
config policy has some templates in the first 2 resources and the
`consolidateManifests` is not set (so it's true) and this should
solve the quay timing issue with the pull secret if using ACM 2.6 or
newer.

Refs:
 - https://github.com/stolostron/backlog/issues/25659
 - https://github.com/stolostron/backlog/issues/25923

Signed-off-by: Gus Parvin <gparvin@redhat.com>