Duplication cleanup (#361)
* Fix typo in policy

The severity field is not defined for Policy - it should be set on the
ConfigurationPolicy instead.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Make placement resources unique

The PlacementBinding and PlacementRule in this policy had the same name
as the PlacementBinding and PlacementRule in another policy. Now those
names match the unique policy name.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Make policy resources unique

The resources in this file had the same names as another (very similar)
file. Now they are uniquely named.

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>
Signed-off-by: Yi Rae Kim <yikim@redhat.com>