ODF 4.19 caused problems with the OPP PolicySet
On OCP 4.19, with the release of ODF 4.19, there is a problem that
prevents OPP from installing successfully.  The ODF operator no longer
provides the StorageSystem CRD.  In ODF 4.19 the customer is expected
to create the StorageCluster without a StorageSystem.

Additionally, the StorageCluster could not be created as configured
in ODF 4.19.  I had to scope the configuration of ODF back to a more
basic install. I have tested the changes on OCP 4.19 and OCP 4.16
successfully.

Refs:
 - https://issues.redhat.com/browse/LPINTEROP-5783

Signed-off-by: Gus Parvin <gparvin@redhat.com>