Changing the ODF configuration for OPP
This is switching the ODF configuration back to the storage device
sets that we were using prior to ODF 4.19.  While this was causing
problems it is the desired configuration from the ODF team and it
appears some fixes on top of ODF 4.19 may have improved the
experience.  I have deployed this successfully on OCP 4.19.

Refs:
 - https://steps.ci.openshift.org/reference/odf-apply-storage-cluster
 - https://issues.redhat.com/browse/LPINTEROP-5791

Signed-off-by: Gus Parvin <gparvin@redhat.com>