Update compliance-operator install config
Previously, the template used `fromClusterClaim`, but if the cluster
claim could not be found, the policy would become noncompliant. Now it
uses `lookup` which will return an empty string if the resource can't be
found. This allows the template to work in more cases.

Users writing their own policy should consider using the newer
`lookupClusterClaim` function, which is easier to read and write. This
policy uses `lookup` directly, for more compatability across versions.

Refs:
 - https://issues.redhat.com/browse/ACM-25409

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>